64bit variables for agents

+++ types.h
@@ -4,7 +4,7 @@
 typedef __m128i *contr;
 typedef uint32_t meter;
 typedef uint16_t place;
-typedef uint32_t agent;
+typedef uint64_t agent;
 typedef uint32_t penny;
 typedef uint32_t edge;
 typedef uint32_t id;
